reduced duplicated code
